+++ src/app/utils/userToken.ts
@@ -1,6 +1,10 @@
+import { JwtPayload } from "jsonwebtoken";
 import { envVars } from "../config/env";
-import { IUser } from "../modules/user/user.interface";
-import { generateToken } from "./jwt";
+import AppError from "../errorHelpers/AppError";
+import { IsActive, IUser } from "../modules/user/user.interface";
+import { User } from "../modules/user/user.model";
+import { generateToken, verifyToken } from "./jwt";
+import httpStatus from 'http-status-codes';
 
 export const createUserToken = (user: Partial<IUser>) => {
     const jwtPayload = {
@@ -16,4 +20,30 @@ export const createUserToken = (user: Partial<IUser>) => {
         accessToken,
         refreshToken
     }
+}
+
+export const createNewAccessTokenWithRefreshToken = async (refreshToken: string) => {
+    const verifiedRefreshToken = verifyToken(refreshToken, envVars.JWT_REFRESH_SECRET) as JwtPayload
+    // we do not have to check the verified status and throw error because if not verified it automatically send error . so no need to write if else 
+
+    const isUserExist = await User.findOne({ email: verifiedRefreshToken.email })
+    if (!isUserExist) {
+        throw new AppError(httpStatus.BAD_REQUEST, "User Does Not Exist")
+    }
+
+    if (isUserExist.isActive === IsActive.BLOCKED || isUserExist.isActive === IsActive.INACTIVE) {
+        throw new AppError(httpStatus.BAD_REQUEST, `User Is ${isUserExist.isActive}`)
+    }
+    if (isUserExist.isDeleted) {
+        throw new AppError(httpStatus.BAD_REQUEST, "User Is Deleted")
+    }
+    // generating access token 
+    const jwtPayload = {
+        userId: isUserExist._id,
+        email: isUserExist.email,
+        role: isUserExist.role
+    }
+    const accessToken = generateToken(jwtPayload, envVars.JWT_ACCESS_SECRET, envVars.JWT_ACCESS_EXPIRES)
+    return accessToken
+
 }
